Ss5

+++ app/Http/Controllers/PostController.php
@@ -4,13 +4,19 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Http\Requests\CreatePostRequest;
+use App\Http\Requests\PostValidationRequest;
 use Illuminate\Support\Facades\Validator;
+use App\Filters\PostFilter;
 use App\Post;
+use App\Notifications\PostValidationNotification;
+use App\PostValidationMessage;
+use Auth;
 
 class PostController extends Controller
 {
     public function __construct(){
-        $this->middleware('auth');
+        $this->middleware('auth')->except(["show",'search','getPosts','getReferences']);
+        $this->middleware('moderator')->only(['validatePost']);
     }
 
 
@@ -60,15 +66,21 @@ class PostController extends Controller
     }
     
     public function attachReference(Post $post, Request $request){
+        if(auth()->user()->id != $post->user_id && !auth()->user()->hasRole(["admin","moderator"])){
+            return abort(404);
+        }
+
         if($request->reference_id !==null){
-            $status = $post->references()->attach($request->reference_id);
-            if($status){
-                return ["status"=>"success"];
-            }
+            $post->references()->attach($request->reference_id);
+            return ["status"=>"success"];
         }
         return ["status"=>"failed"];
     }
     public function dettachReference(Post $post, Request $request){
+        if(auth()->user()->id != $post->user_id && !auth()->user()->hasRole(["admin","moderator"])){
+            return abort(404);
+        }
+        
         if($request->reference_id !==null){
             $status = $post->references()->detach($request->reference_id);
             if($status){
@@ -88,7 +100,7 @@ class PostController extends Controller
     }
 
     public function submitForValidation(Post $post){
-        if($post->status !== "draft"){
+        if($post->status !== "draft" && $post->status !== "rejected"){
             return ["status"=>"faild"];
         }
         $post->status = "validation";
@@ -97,7 +109,101 @@ class PostController extends Controller
     }
 
     public function show(Post $post){
+        if($post->status === "accepted"){
+            return view("post")->with(['post'=>$post]);
+        }
+        if(!Auth::check()){
+            return abort(404);
+        }
+
+        if($post->user_id == auth()->user()->id){
+            return view("post")->with(['post'=>$post]);
+        }
+
+        if($post->status === 'draft'){
+            return abort(404);
+        }
+
+        $user = auth()->user();
+        if($post->status === 'validation' && !$user->hasRole(["moderator","admin"])){
+            return abort(404);
+        }
+
         return view("post")->with(['post'=>$post]);
+    }
+
+    public function search(Request $request){
+        if (empty($request->q)){
+            return redirect('/');
+        }
+        return view('search-result')->with(['searchQuery'=>$request->q]);
+    }
+
+    public function validatePost(Post $post, PostValidationRequest $request){
+        if($post->status === 'draft'){
+            return ['status'=>'failed'];
+        }
+
+        if($request->status === 'rejected'){
+            if(empty($request->message)){
+                return [
+                    'status'=>'failed',
+                    'errors'=>['message required']
+                ];
+            }
+        }
+
+        $postValidationMessage = null;
+        if(!empty($request->message)){
+            $postValidationMessage = new PostValidationMessage;
+            $postValidationMessage->user_id = auth()->id();
+            $postValidationMessage->message = $request->message;
+            $postValidationMessage->status = $request->status;
+            $postValidationMessage->post_id = $post->id;
+            $postValidationMessage->save();
+        }
+
+        
+        $post->status = $request->status;
+        $post->save();
+        $post->user()->first()->notify(new PostValidationNotification($post, $postValidationMessage));
+        if($post->status === "accepted"){
+            $references = $post->references()->get();
+            foreach($references as $reference){
+                $reference->status='accepted';
+                $reference->save();
+            }
+        }
+        return ['status'=>'success'];
+
+    }
+
+    public function getPosts(PostFilter $filter,Request $request){
+        $posts = Post::select('id','title','status','description','user_id','created_at','updated_at')
+        ->filter($filter)->where("status",'accepted');
+        
+        if(Auth::check()){
+            $posts->orWhere([
+                ['status','=','draft'],
+                ['user_id','=',auth()->user()->id],
+            ]);
+
+            if(auth()->user()->hasRole(['admin','moderator'])){
+                $posts->orWhere('status','=','rejected')->orWhere('status','=','validation'); 
+            }
+        }
+        $posts = $posts->paginate(10);
+        $posts->map(function($post){
+            $user=$post->user()->first();
+            $post['username'] = $user === null ?'noname':$user->name;
+        });
+        if(isset($request->desc)){
+            return $posts;
+        }
+        $posts->reverse();
+        return  $posts;
+
+        
     }
 
 }
